refactor: migrate dlna pure parsing to party-display-core
