+++ app/src-tauri/party-display-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod dlna_proxy;
 pub mod system;
 pub mod smtc;
+pub mod dlna;
